+++ ElvUI_EltreumUI/Options/OptionsClassic.lua
@@ -1455,6 +1455,13 @@ if ElvUI_EltreumUI.Classic then
 					icon = 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Icons\\music',
 					order = 85,
 					args = {
+						header1 = {
+							order = 11,
+							type = "description",
+							name = "",
+							width = 'full',
+							image = function() return 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', 3240, 1 end,
+						},
 						moremusicsettings = {
 							order = 1,
 							type = 'group',
@@ -1470,6 +1477,15 @@ if ElvUI_EltreumUI.Classic then
 									get = function(info) return E.private.ElvUI_EltreumUI.combatmusic.enable end,
 									set = function(info, value) E.private.ElvUI_EltreumUI.combatmusic.enable = value E:StaticPopup_Show('PRIVATE_RL') end,
 								},
+								enableboss = {
+									order = 1,
+									type = 'toggle',
+									name = L["Enable"],
+									desc = L["Enable music during combat"],
+									width = 'full',
+									get = function(info) return E.private.ElvUI_EltreumUI.combatmusic.bossmusic end,
+									set = function(info, value) E.private.ElvUI_EltreumUI.combatmusic.bossmusic = value E:StaticPopup_Show('PRIVATE_RL') end,
+								},
 								instancemusic = {
 									order = 2,
 									type = 'toggle',
@@ -1480,16 +1496,23 @@ if ElvUI_EltreumUI.Classic then
 									get = function(info) return E.private.ElvUI_EltreumUI.combatmusic.disableinstance end,
 									set = function(info, value) E.private.ElvUI_EltreumUI.combatmusic.disableinstance = value E:StaticPopup_Show('PRIVATE_RL') end,
 								},
-								pathtofile = {
-									order = 6,
+								header2 = {
+									order = 2,
+									type = "description",
+									name = "Normal Combat Music",
+									width = 'full',
+									image = function() return 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', 3240, 1 end,
+								},
+								combatpathtofile = {
+									order = 3,
 									type = 'group',
 									inline = true,
-									name = L["Path to file"],
+									name = L["Name of file inside Interface\\Addons"],
 									args = {
 										howtomusic = {
 										order = 2,
 										type = "description",
-										name = L["Example: "].."Interface\\Addons\\mymusic.mp3",
+										name = L["Example: "].."mymusic.mp3",
 										},
 										somegap = {
 											order = 3,
@@ -1505,6 +1528,41 @@ if ElvUI_EltreumUI.Classic then
 											width = 'full',
 											get = function() return E.private.ElvUI_EltreumUI.combatmusic.musicfile end,
 											set = function(_, value) E.private.ElvUI_EltreumUI.combatmusic.musicfile = value E:StaticPopup_Show('PRIVATE_RL') end,
+										}
+									}
+								},
+								header3 = {
+									order = 6,
+									type = "description",
+									name = "Boss Combat Music",
+									width = 'full',
+									image = function() return 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', 3240, 1 end,
+								},
+								bosspathtofile = {
+									order = 7,
+									type = 'group',
+									inline = true,
+									name = L["Name of file inside Interface\\Addons"],
+									args = {
+										howtomusic = {
+										order = 2,
+										type = "description",
+										name = L["Example: "].."mymusic.mp3",
+										},
+										somegap = {
+											order = 3,
+											type = "description",
+											name = "",
+										},
+										soundpath = {
+											order = 6,
+											icon = 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Icons\\music',
+											type = 'input',
+											name = '',
+											desc = '',
+											width = 'full',
+											get = function() return E.private.ElvUI_EltreumUI.combatmusic.bossfile end,
+											set = function(_, value) E.private.ElvUI_EltreumUI.combatmusic.bossfile = value E:StaticPopup_Show('PRIVATE_RL') end,
 										}
 									}
 								},